Use 1 URL function from all Remote ViewControllers (+ allow decimal entries in meal view + a couple of small bug fix)

+++ LoopFollow/repository/UserDefaults.swift
@@ -458,7 +458,7 @@ class UserDefaultsRepository {
         static let hideRemoteCustomActions = UserDefaultsValue<Bool>(key: "hideCustomActions", default: true)
     
     //Remote guardrails
-    static let maxCarbs = UserDefaultsValue<Int>(key: "maxCarbs", default: 30)
+    static let maxCarbs = UserDefaultsValue<Double>(key: "maxCarbs", default: 30)
     static let maxBolus = UserDefaultsValue<Double>(key: "maxBolus", default: 2.0)
     
     // API settings
